Move `useTermsSuggestions` to `search-categories` folder
