+++ client/my-sites/plugins/search-categories/index.tsx
@@ -15,10 +15,10 @@ import {
 	useCategories,
 } from 'calypso/my-sites/plugins/categories/use-categories';
 import { useGetCategoryUrl } from 'calypso/my-sites/plugins/categories/use-get-category-url';
-import { useTermsSuggestions } from 'calypso/my-sites/plugins/search-box-header/useTermsSuggestions';
 import { useLocalizedPlugins } from 'calypso/my-sites/plugins/utils';
 import { recordTracksEvent, recordGoogleEvent } from 'calypso/state/analytics/actions';
 import { getSelectedSite } from 'calypso/state/ui/selectors';
+import { useTermsSuggestions } from './useTermsSuggestions';
 
 import './style.scss';
 
@@ -111,8 +111,7 @@ const SearchCategories: FC< {
 		( v ) => [ 'paid', 'popular', 'featured' ].indexOf( v ) < 0
 	);
 	const categories = Object.values( useCategories( displayCategories ) );
-	// console.debug( 'category', category );
-	// console.debug( 'categories', categories );
+
 	// Update the search box with the value from the url everytime it changes
 	// This allows the component to be refilled with a keyword
 	// when navigating back to a page via breadcrumb,
