+++ client/my-sites/plugins/search-categories/useTermsSuggestions.js
